third_party/blink/web_tests/http/tests/devtools/profiler/heap-snapshot-containment-expansion-preserved-when-sorting-expected.txt

Tests Containment view of detailed heap snapshots. Expanded nodes must be preserved after sorting.

Profiler was enabled.
Detailed heap profiles were enabled.

Running: testExpansionPreservedWhenSorting

Profiler was disabled.
